Move the mempool fees function to the rpc module
